smartauth/requirements.txt
asgiref==3.8.1
asyncpg==0.30.0
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
confluent-kafka==2.6.1
constantly==23.10.4
contourpy==1.3.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
cycler==0.12.1
daphne==4.1.2
defusedxml==0.7.1
distlib==0.3.8
dj-rest-auth==7.0.1
Django==5.1.4
django-allauth==65.7.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==25.1
django-formtools==2.5.1
django-jazzmin==3.0.1
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-tinymce==4.1.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filelock==3.13.1
fonttools==4.56.0
gunicorn==23.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
influxdb-client==1.48.0
itypes==1.2.0
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
msgpack==1.1.0
mysql-connector-python==9.2.0
numpy==2.2.3
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
platformdirs==4.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
pyparsing==3.2.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-social-auth==0.3.6
python3-openid==3.2.0
qrcode==7.4.2
reactivex==4.0.4
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
social-auth-core==4.6.1
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.25.0
websockets==14.1
zope.interface==7.2
